+++ mutwo/music_version/__init__.py
@@ -1,4 +1,4 @@
-MAJOR, MINOR, PATCH = 0, 22, 0
+MAJOR, MINOR, PATCH = 0, 23, 0
 
 VERSION = f"{MAJOR}.{MINOR}.{PATCH}"
 """The version of the package ``mutwo.music``."""